ignorar anashei

+++ Pruebas/Rome/Ej4/Ejercicio4.java
@@ -34,6 +34,6 @@ public class Ejercicio4 {
         }
         System.out.println("La cantidad de partidos que gano river fueron: " + ganoRiver);
         System.out.println("La cantidad de goles de Boca siendo local fueron: " + golesBoca);
-        System.out.println("El porcentaje de los partidos que terminaron en empate es: " + ((empate * 100) / 3));
+        System.out.println("El porcentaje de los partidos que terminaron en empate es: " + ((double) (empate * 100) / 3));
     }
 }
